# Fixing the issue of wrong get value