@soniabhishek 2018-10-23
@buger can you review this PR asap This is fixing the case I was facing the issue if you think the approach is correct, then I will add few test cases and we can merge this one.